+++ src/llmq/quorums_dkgsessionmgr.cpp
@@ -94,7 +94,7 @@ void CDKGSessionManager::ProcessMessage(CNode* pfrom, const std::string& strComm
     }
 
     // peek into the message and see which uint8_t it is. First byte of all messages is always the uint8_t
-    uint8_t llmqType = (uint8_t)*vRecv.begin();
+    uint8_t llmqType = *vRecv.begin();
     if (!dkgSessionHandlers.count(llmqType)) {
         LOCK(cs_main);
         Misbehaving(pfrom->GetId(), 100);
